Rename SimpleSidebar.tsx to Sidebar.tsx
- Rename SimpleSidebar component to standard Sidebar name
- Update exports in index.ts to reference correct filename
- Maintain clean naming convention for component files
