+++ frontend/src/shared/components/Sidebar.tsx
